requirements.txt
backcall==0.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
colorama==0.4.5
decorator==5.1.1
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
lxml==4.9.2
numpy==1.19.5
parso==0.7.1
pickleshare==0.7.5
praat-parselmouth==0.4.3
praat-textgrids==1.4.0
prompt-toolkit==3.0.36
Pygments==2.14.0
six==1.16.0
soupsieve==2.3.2.post1
TextGrid==1.5
tgt==1.4.4
traitlets==4.3.3
wcwidth==0.2.6
